+++ core/src/main/java/com/elepy/RouteGenerator.java
@@ -5,9 +5,9 @@ import com.elepy.describers.ClassDescriber;
 import com.elepy.describers.ModelDescription;
 import com.elepy.describers.ResourceDescriber;
 import com.elepy.exceptions.ElepyConfigException;
+import com.elepy.http.AccessLevel;
 import com.elepy.http.HttpContext;
 import com.elepy.http.HttpMethod;
-import com.elepy.models.AccessLevel;
 import com.elepy.utils.ClassUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
