+++ core/src/main/java/com/elepy/annotations/Create.java
@@ -1,7 +1,7 @@
 package com.elepy.annotations;
 
 
-import com.elepy.models.AccessLevel;
+import com.elepy.http.AccessLevel;
 import com.elepy.routes.CreateHandler;
 import com.elepy.routes.DefaultCreate;
