+++ core/src/main/java/com/elepy/annotations/Find.java
@@ -1,7 +1,7 @@
 package com.elepy.annotations;
 
 
-import com.elepy.models.AccessLevel;
+import com.elepy.http.AccessLevel;
 import com.elepy.routes.DefaultFind;
 import com.elepy.routes.FindHandler;
