+++ core/src/main/java/com/elepy/annotations/Route.java
@@ -1,7 +1,7 @@
 package com.elepy.annotations;
 
+import com.elepy.http.AccessLevel;
 import com.elepy.http.HttpMethod;
-import com.elepy.models.AccessLevel;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
