+++ core/src/main/java/com/elepy/annotations/Update.java
@@ -1,7 +1,7 @@
 package com.elepy.annotations;
 
 
-import com.elepy.models.AccessLevel;
+import com.elepy.http.AccessLevel;
 import com.elepy.routes.DefaultUpdate;
 import com.elepy.routes.UpdateHandler;
